predictions/requirements.txt
## Requirements.txt file

# Data manipulation and analysis
numpy==1.24.4
pandas==2.0.3

# Data visualisation
matplotlib==3.4.3
seaborn==0.12.2
statsmodels==0.14.0

# Machine Learning
scikit-learn==1.3.0

# Filesystem handling 
pathlib==1.0.1
python-dotenv==1.0.0

# Model Tracking
mlflow==2.5.0

# AWS SDK for Python - used for interacting with AWS services
boto3==1.28.2

# PostgreSQL 
psycopg==3.1.9
psycopg_binary==3.1.9
psycopg2_binary==2.9.6

# Prefect
prefect==2.10.21
prefect-aws==0.3.5

# Deployment
flask==2.3.2
gunicorn>=20,<22
requests==2.28.1

# Testing
pytest==7.4.0
pylint==2.17.4
psutil==5.9.5